requirements.txt
Django==1.11.4
djangorestframework==3.6.4
Pygments==2.2.0
pytz==2017.2
